# api/openapi.yaml
openapi: "3.0.3"
info:
  title: NuoNetDisk API
  description: Self-hosted cloud storage API
  version: "0.1.0"
  contact:
    name: NuoNetDisk Team
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: NOT_FOUND
            message:
              type: string
              description: Human-readable error message
              example: The requested resource was not found
...
    AuthRefreshRequest:
      type: object
      required:
        - refresh_token
      properties:
        refresh_token:
          type: string
    AuthTokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          description: Access token TTL in seconds
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        size:
          type: integer
          format: int64
        mime_type:
          type: string
        sha256_hash:
          type: string
        parent_folder_id:
          type: string
          format: uuid
          nullable: true
        is_deleted:
          type: boolean
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
    Folder:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        name:
          type: string
        parent_folder_id:
          type: string
          format: uuid
          nullable: true
        is_deleted:
          type: boolean
        version:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
        expires_at:
          type: string
          format: date-time
          nullable: true
...
    PaginatedFiles:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/File"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean
    PaginatedFolders:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Folder"
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        has_more:
          type: boolean
...
paths:
...
  /auth/refresh:
    post:
      tags:
        - Auth
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRefreshRequest"
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthTokenResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /recycle-bin/files:
    get:
      tags:
        - RecycleBin
      summary: List soft-deleted files
      operationId: listDeletedFiles
      security:
        - BearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
            minimum: 1
      responses:
        "200":
          description: Paginated deleted files
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFiles"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recycle-bin/folders:
    get:
      tags:
        - RecycleBin
      summary: List soft-deleted folders
      operationId: listDeletedFolders
      security:
        - BearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
            minimum: 1
      responses:
        "200":
          description: Paginated deleted folders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedFolders"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recycle-bin/restore/file/{fileId}:
    post:
      tags:
        - RecycleBin
      summary: Restore a soft-deleted file
      operationId: restoreFile
      security:
        - BearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: File restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/File"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found in recycle bin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recycle-bin/restore/folder/{folderId}:
    post:
      tags:
        - RecycleBin
      summary: Restore a soft-deleted folder
      operationId: restoreFolder
      security:
        - BearerAuth: []
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Folder and contents restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Folder"
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Folder not found in recycle bin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /recycle-bin/file/{fileId}:
    delete:
      tags:
        - RecycleBin
      summary: Permanently delete a file
      operationId: permanentDeleteFile
      security:
        - BearerAuth: []
      parameters:
        - name: fileId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: File permanently deleted
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: File not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      tags:
        - Admin
      summary: List all users (admin only)
      operationId: listUsers
      security:
        - BearerAuth: []
      parameters:
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: date_from
          in: query
          description: Inclusive lower bound on user created_at (ISO date, YYYY-MM-DD).
          schema:
            type: string
            format: date
        - name: date_to
          in: query
          description: Inclusive upper bound on user created_at (ISO date, YYYY-MM-DD).
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Paginated user list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                        email:
                          type: string
                        display_name:
                          type: string
                        avatar_url:
                          type: string
                          description: URL path to fetch the user's avatar (admin-only endpoint), empty if no avatar set.
                        bio:
                          type: string
                        gender:
                          type: string
                        is_admin:
                          type: boolean
                        is_super_admin:
                          type: boolean
                          description: True if this user is the configured super admin (env ADMIN_EMAIL).
                        created_at:
                          type: string
                          format: date-time
                        updated_at:
                          type: string
                          format: date-time
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
                  has_more:
                    type: boolean
        "403":
          description: Forbidden
...
  /admin/users/register:
    post:
      tags:
        - Admin
      summary: Create a new admin account (super admin only)
      description: |
        Only the super admin account (configured via ADMIN_EMAIL) can create
        new admin accounts. Regular admins receive a 403.
      operationId: registerAdmin
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 8
      responses:
        "201":
          description: Admin account created
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden - super admin access required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /recycle-bin/folder/{folderId}:
    delete:
      tags:
        - RecycleBin
      summary: Permanently delete a folder
      operationId: permanentDeleteFolder
      security:
        - BearerAuth: []
      parameters:
        - name: folderId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Folder and contents permanently deleted
        "401":
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Folder not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"